Fixes #18: Preload images.
Preloads the next and previous image when a user navigates to an image.
Preloading happens right after the navigation is done.

Calling load on an already loaded image will not cause the browser to
do another request, this way the method can be called multiple times
without worry.

Uses Promise.all to make sure both images get loaded as fast as
possible.